backend/requirements/base.txt
# Base requirements for Roska Radiadores backend

# Django
Django>=5.0,<5.1
django-environ>=0.11.2
python-decouple>=3.8

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7

# Database
psycopg2-binary>=2.9.9

# CORS
django-cors-headers>=4.3.1

# Filtering
django-filter>=23.5

# File handling
Pillow>=10.2.0

# Cerbos SDK
cerbos>=0.9.0

# Celery (for async tasks)
celery>=5.3.4
redis>=5.0.1

# Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
